requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coverage==6.3.3
Deprecated==1.2.13
Faker==13.11.1
fastapi==0.71.0
fastapi-jwt-auth==0.5.0
flake8==4.0.1
h11==0.13.0
hypothesis==6.46.3
idna==3.3
iniconfig==1.1.1
jdatetime==4.1.0

mccabe==0.6.1

packaging==21.3
passlib==1.7.4
pika==1.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==1.7.1
pymongo==4.0.1
pyparsing==3.0.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.19.2
redis==4.1.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
wrapt==1.13.3

